refactor: make i18n pkg
