+++ packages/i18n/jest.config.js
@@ -1,0 +1,18 @@
+export default {
+  displayName: '@europeana/i18n',
+  moduleFileExtensions: [
+    'js',
+    'json',
+    'mjs'
+  ],
+  moduleNameMapper: {
+    '^@/(.*)$': '<rootDir>/src/$1'
+  },
+  testEnvironment: 'jsdom',
+  testPathIgnorePatterns: [
+    '/node_modules/'
+  ],
+  transform: {
+    '^.+\\.(js|mjs)$': 'babel-jest'
+  }
+};
