+++ packages/i18n/src/index.js
@@ -1,0 +1,9 @@
+import codes from './codes.js';
+import langMap from './langMap.js';
+import locales from './locales.js';
+
+export default {
+  codes,
+  langMap,
+  locales
+};
