+++ packages/i18n/src/locales.js
@@ -2,175 +2,150 @@ export default [
   {
     name: 'Български',
     code: 'bg',
-    file: 'bg.js',
     iso: 'bg-BG',
     isoAlpha3: 'bgr'
   },
   {
     name: 'Čeština',
     code: 'cs',
-    file: 'cs.js',
     iso: 'cs-CZ',
     isoAlpha3: 'cze'
   },
   {
     name: 'Dansk',
     code: 'da',
-    file: 'da.js',
     iso: 'da-DK',
     isoAlpha3: 'dnk'
   },
   {
     name: 'Deutsch',
     code: 'de',
-    file: 'de.js',
     iso: 'de-DE',
     isoAlpha3: 'deu'
   },
   {
     name: 'Ελληνικά',
     code: 'el',
-    file: 'el.js',
     iso: 'el-GR',
     isoAlpha3: 'grc'
   },
   {
     name: 'English',
     code: 'en',
-    file: 'en.js',
     iso: 'en-GB',
     isoAlpha3: 'eng'
   },
   {
     name: 'Español',
     code: 'es',
-    file: 'es.js',
     iso: 'es-ES',
     isoAlpha3: 'esp'
   },
   {
     name: 'Eesti',
     code: 'et',
-    file: 'et.js',
     iso: 'et-EE',
     isoAlpha3: 'est'
   },
   {
     name: 'Euskara',
     code: 'eu',
-    file: 'eu.js',
     iso: 'eu-ES',
     isoAlpha3: 'eus'
   },
   {
     name: 'Suomi',
     code: 'fi',
-    file: 'fi.js',
     iso: 'fi-FI',
     isoAlpha3: 'fin'
   },
   {
     name: 'Français',
     code: 'fr',
-    file: 'fr.js',
     iso: 'fr-FR',
     isoAlpha3: 'fra'
   },
   {
     name: 'Gaeilge',
     code: 'ga',
-    file: 'ga.js',
     iso: 'ga-IE',
     isoAlpha3: 'gle'
   },
   {
     name: 'Hrvatski',
     code: 'hr',
-    file: 'hr.js',
     iso: 'hr-HR',
     isoAlpha3: 'hrv'
   },
   {
     name: 'Magyar',
     code: 'hu',
-    file: 'hu.js',
     iso: 'hu-HU',
     isoAlpha3: 'hun'
   },
   {
     name: 'Italiano',
     code: 'it',
-    file: 'it.js',
     iso: 'it-IT',
     isoAlpha3: 'ita'
   },
   {
     name: 'Lietuvių',
     code: 'lt',
-    file: 'lt.js',
     iso: 'lt-LT',
     isoAlpha3: 'lit'
   },
   {
     name: 'Latviešu',
     code: 'lv',
-    file: 'lv.js',
     iso: 'lv-LV',
     isoAlpha3: 'lav'
   },
   {
     name: 'Malti',
     code: 'mt',
-    file: 'mt.js',
     iso: 'mt-MT',
     isoAlpha3: 'mlt'
   },
   {
     name: 'Nederlands',
     code: 'nl',
-    file: 'nl.js',
     iso: 'nl-NL',
     isoAlpha3: 'nld'
   },
   {
     name: 'Polski',
     code: 'pl',
-    file: 'pl.js',
     iso: 'pl-PL',
     isoAlpha3: 'pol'
   },
   {
     name: 'Português',
     code: 'pt',
-    file: 'pt.js',
     iso: 'pt-PT',
     isoAlpha3: 'por'
   },
   {
     name: 'Română',
     code: 'ro',
-    file: 'ro.js',
     iso: 'ro-RO',
     isoAlpha3: 'ron'
   },
   {
     name: 'Slovenčina',
     code: 'sk',
-    file: 'sk.js',
     iso: 'sk-SK',
     isoAlpha3: 'slk'
   },
   {
     name: 'Slovenščina',
     code: 'sl',
-    file: 'sl.js',
     iso: 'sl-SI',
     isoAlpha3: 'slv'
   },
   {
     name: 'Svenska',
     code: 'sv',
-    file: 'sv.js',
     iso: 'sv-SE',
     isoAlpha3: 'swe'
   }
