+++ packages/portal/Dockerfile
@@ -12,6 +12,7 @@ RUN mkdir -p packages/portal
 
 COPY package*.json ./
 COPY packages/portal/package*.json ./packages/portal/
+COPY packages/i18n ./packages/i18n
 COPY packages/mirador ./packages/mirador
 COPY packages/oembed ./packages/oembed
 COPY packages/style ./packages/style
