+++ packages/portal/package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@elastic/apm-rum-vue": "^1.3.1",
+    "@europeana/i18n": "^1.149.4",
     "@europeana/oembed": "^1.148.0",
     "@europeana/vue-session": "^1.132.0",
     "@europeana/vue-router-query": "^1.148.0",
