+++ packages/portal/src/cachers/utils.js
@@ -1,7 +1,8 @@
 import axios from 'axios';
 import { createClient } from 'redis';
 import _pick from 'lodash/pick.js';
-import { daily, langMapValueForLocale } from '../plugins/europeana/utils.js';
+import { daily } from '../plugins/europeana/utils.js';
+import { langMapValueForLocale } from '@europeana/i18n/langMap.js';
 import EuropeanaRecordApi from '../plugins/europeana/record.js';
 import EuropeanaEntityApi from '../plugins/europeana/entity.js';
