+++ packages/portal/src/components/item/ItemPinButton.vue
@@ -49,7 +49,7 @@
 <script>
   import makeToastMixin from '@/mixins/makeToast';
   import entityBestItemsSetMixin from '@/mixins/europeana/entities/entityBestItemsSet';
-  import { langMapValueForLocale } from '@/plugins/europeana/utils';
+  import { langMapValueForLocale } from '@europeana/i18n/langMap.js';
 
   export default {
     name: 'ItemPinButton',
