+++ packages/portal/src/components/item/ItemPinModal.vue
@@ -74,7 +74,7 @@
   import entityBestItemsSetMixin from '@/mixins/europeana/entities/entityBestItemsSet';
   import langAttributeMixin from '@/mixins/langAttribute';
   import makeToastMixin from '@/mixins/makeToast';
-  import { langMapValueForLocale } from '@/plugins/europeana/utils';
+  import { langMapValueForLocale } from '@europeana/i18n/langMap.js';
 
   export default {
     name: 'ItemPinModal',
