+++ packages/portal/src/components/item/ItemPreviewCard.vue
@@ -70,7 +70,7 @@
 </template>
 
 <script>
-  import { langMapValueForLocale } from  '@/plugins/europeana/utils';
+  import { langMapValueForLocale } from '@europeana/i18n/langMap.js';
 
   import ContentCard from '../content/ContentCard';
