+++ packages/portal/src/components/set/SetAddItemButton.vue
@@ -15,7 +15,7 @@
 </template>
 
 <script>
-  import { langMapValueForLocale } from '@/plugins/europeana/utils';
+  import { langMapValueForLocale } from '@europeana/i18n/langMap.js';
 
   export default {
     name: 'SetAddItemButton',
