+++ packages/portal/src/components/set/SetCardGroup.vue
@@ -29,7 +29,7 @@
 </template>
 
 <script>
-  import { getLabelledSlug } from '@/plugins/europeana/utils';
+  import { getLabelledSlug } from '@/plugins/europeana/utils.js';
 
   export default {
     name: 'SetCardGroup',
