+++ packages/portal/src/components/set/SetPublicationRequestWidget.vue
@@ -103,7 +103,7 @@
 </template>
 
 <script>
-  import { langMapValueForLocale } from  '@/plugins/europeana/utils';
+  import { langMapValueForLocale } from '@europeana/i18n/langMap.js';
   import axios from 'axios';
 
   export default {
