+++ packages/portal/src/mixins/collectionLinkGen.js
@@ -1,5 +1,5 @@
 import { getEntityTypeHumanReadable  } from '@/plugins/europeana/entity';
-import { getLabelledSlug } from '@/plugins/europeana/utils';
+import { getLabelledSlug } from '@/plugins/europeana/utils.js';
 import { BASE_URL as EUROPEANA_DATA_URL } from '@/plugins/europeana/data';
 
 export default {
