+++ packages/portal/src/pages/contentful/entity-harvester/index.vue
@@ -31,7 +31,8 @@
     entityParamsFromUri
   } from '@/plugins/europeana/entity';
   import contentfulSidebarMixin from '@/mixins/contentful/sidebar';
-  import { langMapValueForLocale, getLabelledSlug } from '@/plugins/europeana/utils';
+  import { getLabelledSlug } from '@/plugins/europeana/utils.js';
+  import { langMapValueForLocale } from '@europeana/i18n/langMap.js';
   import { BASE_URL } from '@/plugins/europeana/data';
 
   export default {
